feature: new project structure
