+++ generative_mps/data/parameters.yaml
@@ -1,0 +1,15 @@
+output_dir: windowed_ti/real
+training_image: strebelle.png
+checkpoint: .checkpoints/
+sample_images: output/wgan
+
+num_channels: 1
+latent_dim: 100
+learning_rate: 0.0002
+images_path: windowed_ti/
+batch_size: 64
+num_workers: 8
+
+n_epochs: 100
+cuda: True
+n_critic: 5
