--- src/helpers/train_utils.py
@@ -1,65 +1,0 @@
-"""
-Helper module with useful functions for training stage
-"""
-
-import os
-import torch
-from torch.autograd import grad
-from torch.autograd import Variable
-import shutil
-
-def gradient_penalty(x,y,f):
-    shape =[x.size(0)] + [1] * (x.dim() -1)
-    alpha = torch.rand(shape)
-    z = x+ alpha *(y-x)
-    z = Variable(z,requires_grad=True)
-    z=z
-    o=f(z)
-    g = grad(o,z, grad_outputs=torch.ones(o.size()), create_graph=True)[0].view(z.size(0), -1)
-    gp = ((g.norm(p=2, dim=1))**2).mean()
-    return gp
-
-
-def save_checkpoint(state, save_path, is_best=False, max_keep=None):
-    # save checkpoint
-    torch.save(state, save_path)
-
-    # deal with max_keep
-    save_dir = os.path.dirname(save_path)
-    list_path = os.path.join(save_dir, 'latest_checkpoint')
-
-    save_path = os.path.basename(save_path)
-    if os.path.exists(list_path):
-        with open(list_path) as f:
-            ckpt_list = f.readlines()
-            ckpt_list = [save_path + '\n'] + ckpt_list
-    else:
-        ckpt_list = [save_path + '\n']
-
-    if max_keep is not None:
-        for ckpt in ckpt_list[max_keep:]:
-            ckpt = os.path.join(save_dir, ckpt[:-1])
-            if os.path.exists(ckpt):
-                os.remove(ckpt)
-        ckpt_list[max_keep:] = []
-
-    with open(list_path, 'w') as f:
-        f.writelines(ckpt_list)
-
-    # copy best
-    if is_best:
-        shutil.copyfile(save_path, os.path.join(save_dir, 'best_model.ckpt'))
-
-
-def load_checkpoint(ckpt_dir_or_file, map_location=None, load_best=False):
-    if os.path.isdir(ckpt_dir_or_file):
-        if load_best:
-            ckpt_path = os.path.join(ckpt_dir_or_file, 'best_model.ckpt')
-        else:
-            with open(os.path.join(ckpt_dir_or_file, 'latest_checkpoint')) as f:
-                ckpt_path = os.path.join(ckpt_dir_or_file, f.readline()[:-1])
-    else:
-        ckpt_path = ckpt_dir_or_file
-    ckpt = torch.load(ckpt_path, map_location=map_location)
-    print('[INFO] Loading checkpoint from %s succeed!' % ckpt_path)
-    return ckpt
